+++ examples/depth.rs
@@ -197,8 +197,15 @@ fn main() {
                 }
                 brush.queue(&section);
                 brush.queue(&section2);
+                match brush.process_queued(&device, &queue) {
+                    Ok(_) => (),
+                    Err(err) => {
+                        eprintln!("{err}");
+                        panic!("");
+                    }
+                };
 
-                let cmd_buffer = brush.draw_with_depth(&device, &queue, &view).unwrap();
+                let cmd_buffer = brush.draw_with_depth(&device, &view);
 
                 // Has to be submitted last so it won't be overlapped.
                 queue.submit([encoder.finish(), cmd_buffer]);